Partial Code segmentation done🤘
